notebooklm/notebooklm-backend/requirements.txt
# Core Flask dependencies
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0

# Database
SQLAlchemy==2.0.23

# Environment and configuration
python-dotenv==1.0.0

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23
openpyxl==3.1.2
markdown==3.5.1
beautifulsoup4==4.12.2
lxml==4.9.3

# AI and ML
openai[datalib]>=1.35.13
anthropic==0.7.7
google-generativeai==0.3.2
cohere==4.37

# Vector database and embeddings
chromadb==0.4.18
sentence-transformers
numpy==1.24.3
scikit-learn==1.3.2

# Text processing
nltk==3.8.1
spacy==3.7.2
tiktoken==0.5.2

# Audio processing and TTS
pydub==0.25.1
elevenlabs==0.2.26
gTTS==2.4.0
openai-whisper==20231117
SpeechRecognition==3.10.0

# HTTP requests
requests==2.31.0
httpx==0.25.2

# Utilities
python-magic==0.4.27
Pillow==10.1.0
python-dateutil==2.8.2
pytz==2023.3

# Development and testing
pytest==7.4.3
pytest-flask==1.3.0
black==23.11.0
flake8==6.1.0

# Production server
gunicorn==21.2.0
gevent==23.9.1

# Monitoring and logging
structlog==23.2.0

# YouTube integration
youtube_transcript_api==0.6.2

